@@ -1,6 +1,6 @@
 {
     "name": "@bacchilu/react-bootstrap-modal",
-    "version": "1.1.1",
+    "version": "2.0.0",
     "description": "ReactJS Bootstrap Modal",
     "keywords": [
         "react",
@@ -22,17 +22,17 @@
     "author": "Luca Bacchi <bacchilu@gmail.com> (https://github.com/bacchilu)",
     "license": "ISC",
     "devDependencies": {
-        "@babel/core": "^7.17.5",
-        "@babel/preset-env": "^7.16.11",
+        "@babel/core": "^7.17.10",
+        "@babel/preset-env": "^7.17.10",
         "@babel/preset-react": "^7.16.7",
-        "babel-loader": "^8.2.3",
-        "webpack": "^5.70.0",
+        "babel-loader": "^8.2.5",
+        "webpack": "^5.72.0",
         "webpack-cli": "^4.9.2",
-        "npm-check-updates": "^12.5.3"
+        "npm-check-updates": "^12.5.11"
     },
     "dependencies": {
-        "core-js": "^3.21.1",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "core-js": "^3.22.4",
+        "react": "^18.1.0",
+        "react-dom": "^18.1.0"
     }
 }
